# Custom integer64 slicing support