--- a/src/slice-chop.c
+++ b/src/slice-chop.c
@@ -322,7 +322,9 @@ static SEXP chop_fallback(SEXP x, SEXP indices, struct vctrs_chop_info info) {
   Rf_defineVar(syms_x, x, env);
   Rf_defineVar(syms_i, info.index, env);
 
-  // Construct call with symbols, not values, for performance
+  // Construct call with symbols, not values, for performance.
+  // TODO - Remove once bit64 is updated on CRAN. Special casing integer64
+  // objects to ensure correct slicing with `NA_integer_`.
   SEXP call;
   if (is_integer64(x)) {
     call = PROTECT(Rf_lang3(syms_vec_slice_dispatch_integer64, syms_x, syms_i));