--- a/src/slice.h
+++ b/src/slice.h
@@ -1,6 +1,8 @@
 #ifndef VCTRS_SLICE_H
 #define VCTRS_SLICE_H
 
+extern SEXP syms_vec_slice_dispatch_integer64;
+extern SEXP fns_vec_slice_dispatch_integer64;
 
 SEXP slice_names(SEXP names, SEXP subscript);
 SEXP slice_rownames(SEXP names, SEXP subscript);