Closes tidyverse/tidyr#846 

This PR adds integer64 support for `vec_chop()` and `vec_slice()`, repairing the object whenever it is sliced with `NA_integer_`, which is unsupported by the cran version of bit64.

By extension this allows us to use `vec_init()` with these objects, freeing us to hard-code casting of `unspecified()` objects to any `to` type with `vec_init(to, vec_size(x))` in #812 

This adds two new R functions:

- `vec_slice_fallback_integer64()` for use with shaped integer64 objects
- `vec_slice_dispatch_integer64()` for use with 1D integer64 objects

When slicing with shaped objects, the C level `vec_slice_fallback()` will switch to calling the R level `vec_slice_fallback_integer64()` if the input is integer64.

When slicing 1D objects with `[`, I've added a new C level `vec_slice_dispatch()` that either calls out to `vec_slice_dispatch_integer64()` or does our original behavior of calling to `[`.

The rationale for doing the `inherits(x, "integer64")` check at the C level is for speed with `vec_chop()`. We don't want to slow down `vec_chop()` on things like factors by making it repeatedly call `inherits(x, "integer64")` to decide which slicing path to take. We can determine whether we need `[` or `vec_slice_dispatch_integer64()` once up front, and then construct the slice call using that function. It feels a little invasive because now `chop_fallback()` has to "know" about integer64 objects, but the benefits seem worth it

This makes the assumption that if `inherits(x, "integer64")` is true, then the user has bit64 installed. I can add an `is_installed("bit64")` check to `vec_slice_fallback_integer64()` and `vec_slice_dispatch_integer64()` if you feel it is worth it.