@DavisVaughan 2020-02-13
Some proof that I am not wrecking `vec_slice()` and `vec_chop()` performance with factors

Large factor:

```r
library(vctrs)
set.seed(123)

x <- factor(sample(letters, 1e4, replace = TRUE))
idx <- 1:100 + 0L

# before
bench::mark(vec_slice(x, idx), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_slice(x, idx)   7.72µs   11.4µs    78034.    17.6KB     72.6

# after
bench::mark(vec_slice(x, idx), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_slice(x, idx)    7.7µs   10.9µs    82997.    17.6KB     77.3



# before
bench::mark(vec_chop(x), iterations = 1000)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_chop(x)   33.3ms   90.9ms      11.8    82.2KB     21.5

# after
bench::mark(vec_chop(x), iterations = 1000)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_chop(x)   32.5ms   90.7ms      11.8    82.2KB     21.6
```

Small factor:

``` r
library(vctrs)
set.seed(123)

x <- factor(c("a", "b", "c"))
idx <- 1:2 + 0L

# before
bench::mark(vec_slice(x, idx), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_slice(x, idx)   6.83µs   9.62µs    93748.    17.1KB     87.3

# after
bench::mark(vec_slice(x, idx), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_slice(x, idx)   6.49µs   9.19µs    99387.    17.1KB     92.5



# before
bench::mark(vec_chop(x), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_chop(x)   11.1µs   13.6µs    70837.    4.01KB     29.8

# after
bench::mark(vec_chop(x), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_chop(x)   11.6µs   14.1µs    69094.    4.01KB     29.0
```