@lionel- 2020-02-13
Can you add a NEWS item please? Also check if that tidyverse/tidyr#846 is sloved and close it from here in this case.