@DavisVaughan 2020-02-13
Closes tidyverse/tidyr#846 

Confirmed!

``` r
library(tidyr)
library(vctrs)
library(bit64)

fish_encounters$seen <- as.integer64(fish_encounters$seen)
fish_encounters <- fish_encounters[1:30,]

# CRAN
fish_encounters %>%
  pivot_wider(names_from = station, values_from = seen)
#> # A tibble: 3 x 12
#>   fish  Release I80_1  Lisbon  Rstr  Base_TD BCE   BCW   BCE2  BCW2  MAE   MAW  
#>   <fct> <int64> <int6> <int64> <int> <int64> <int> <int> <int> <int> <int> <int>
#> 1 4842  1       1      1       1     1       1     1     1         …     …     …
#> 2 4843  1       1      1       1     1       1     1     1         …     …     …
#> 3 4844  1       1      1       1     1       1     1     1     9218… 9218… 9218…

# This PR
fish_encounters %>%
  pivot_wider(names_from = station, values_from = seen)
#> # A tibble: 3 x 12
#>   fish  Release I80_1  Lisbon  Rstr  Base_TD BCE   BCW   BCE2  BCW2  MAE   MAW  
#>   <fct> <int64> <int6> <int64> <int> <int64> <int> <int> <int> <int> <int> <int>
#> 1 4842  1       1      1       1     1       1     1     1      1     1     1   
#> 2 4843  1       1      1       1     1       1     1     1      1     1     1   
#> 3 4844  1       1      1       1     1       1     1     1     NA    NA    NA
```

<sup>Created on 2020-02-13 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>